--- conf/Makefile/Linux_Fedora29.mk
@@ -1,1 +1,0 @@
-Linux_Fedora27.mk
