+++ src/uninstall.py
@@ -9,16 +9,16 @@ import install
 import util
 
 if util.on_macos:
-    import OSes.darwin as OS
+    import OSes.macos as OS
 
 elif util.on_wsl:
     import OSes.wsl as OS
 
 elif util.on_linux:
     if config.NAME == 'Ubuntu':
-        import OSes.debian as OS
-    else:
-        import OSes.redhat as OS
+        import OSes.ubuntu as OS
+    #else:
+    #    import OSes.debian as OS
 
 
 def main(name=config.DOCKER_CONTAINER_NAME, tag=config.DOCKER_CONTAINER_TAG, tld=config.TOP_LEVEL_DOMAIN):
