--- supported_os.json
@@ -1,18 +1,0 @@
-{
-    "Ubuntu": {
-        "min": 16,
-        "max": 20
-    },
-    "Fedora": {
-        "min": 27,
-        "max": 31
-    },
-    "macOS": {
-        "min": "10.12",
-        "max": "10.19"
-    },
-    "wsl2": {
-        "min": 18,
-        "max": 18
-    }
-}
